+++ CHANGES.md
@@ -48,6 +48,7 @@
 -   pat date picker: Support updating a date if it is before another dependent date.
 -   pat tabs: Refactor based on ``ResizeObserver`` and fix problems calculating the with with transitions.
 -   pat tabs: When clicking on the ``extra-tabs`` element, toggle between ``open`` and ``closed`` classes to allow opening/closing an extra-tabs menu via CSS.
+-   pat autofocus: Do not autofocus in iframes. Fixes: #761.
 
 ### Technical